+++ include/notcurses/notcurses.h
@@ -675,7 +675,8 @@ cell_strdup(const struct ncplane* n, const cell* c){
 
 // Extract the three elements of a cell.
 static inline char*
-cell_extract(const struct ncplane* n, const cell* c, uint32_t* attrword, uint64_t* channels){
+cell_extract(const struct ncplane* n, const cell* c,
+             uint32_t* attrword, uint64_t* channels){
   if(attrword){
     *attrword = c->attrword;
   }
@@ -1166,7 +1167,7 @@ ncplane_at_cursor_cell(struct ncplane* n, cell* c){
 // Retrieve the current contents of the specified cell. The EGC is returned, or
 // NULL on error. This EGC must be free()d by the caller. The attrword and
 // channels are written to 'attrword' and 'channels', respectively.
-API char* ncplane_at_yx(struct ncplane* n, int y, int x,
+API char* ncplane_at_yx(const struct ncplane* n, int y, int x,
                         uint32_t* attrword, uint64_t* channels);
 
 // Retrieve the current contents of the specified cell into 'c'. This cell is